% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_id_columns.R
\name{add_group_id_columns}
\alias{add_group_id_columns}
\title{Add group membership column for each group}
\usage{
add_group_id_columns(tbl_x, cell_name_col, group_col)
}
\arguments{
\item{tbl_x}{A tibble containing a cell name column and a group column}

\item{cell_name_col}{The name of the column that contains the cell identifier}

\item{group_col}{The column identifying group membership}
}
\description{
This function is mainly designed to support plotting of cell membership for individual groups, for example individual clusters
}
\examples{
library(mkhelpers)
library(Seurat)
library(SeuratData)
InstallData("pbmc3k")
seu_pbmc <- LoadData("pbmc3k", "pbmc3k.final")
tbl_x <- seu_extract_tbl(seu_x = seu_pbmc,
                         reduction = "umap",
                         metadata_cols = "seurat_clusters")
tbl_x <- tbl_x \%>\% add_group_id_columns(cell_name_col = cell_name, group_col = seurat_clusters)
show_labels_list <- c(TRUE, rep(FALSE, length(tbl_x$seurat_clusters \%>\% unique)))
plot_umap_grid(tbl_x = tbl_x, umap_dim_col_1 = UMAP_1, umap_dim_col_2 = UMAP_2, output_path = "~/test_plot_cluster_ids.png", feature_list = str_subset(colnames(tbl_x), "^seurat_clusters"), show_labels = show_labels_list, point_size = 0.5, show_legend = FALSE)
}
