% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_quantile_cutoffs.R
\name{apply_quantile_cutoffs}
\alias{apply_quantile_cutoffs}
\title{Applying quantile cutoffs to a numeric vector}
\usage{
apply_quantile_cutoffs(x, quantiles = c(0.1, 0.9))
}
\arguments{
\item{x}{A numeric vector}

\item{quantiles}{Lower and upper quantiles to apply}
}
\value{
A numeric vector
}
\description{
Elements in the vector <= the lower cutoff will be set to the lower cutoff.

Elements in the vector >= the upper cutoff will be set to the upper cutoff.

All other elements will be left unchanged
}
\examples{
library(mkhelpers)
mtcars$disp \%>\% sort
apply_quantile_cutoffs(mtcars$disp) \%>\% sort
}
