% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bro_ggsave_paged.R
\name{bro_ggsave_paged}
\alias{bro_ggsave_paged}
\title{Universal ggplot export function from github.com/jbengler/bro with minor fix for current ggplot versions}
\usage{
bro_ggsave_paged(
  gg = ggplot2::last_plot(),
  filename,
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = "mm",
  dpi = 300,
  limitsize = TRUE,
  return_input = FALSE,
  burst_to_multiple_files = FALSE,
  ...
)
}
\arguments{
\item{gg}{ggplot or list of ggplots}

\item{filename}{Filename for export}

\item{device}{Device to use. Can either be a device function (e.g. `png()`), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in `units` ("mm", "cm", or "in").
If not supplied, uses the size of current graphics device.
In case the input has absolute dimensions after applying
`patchwork::plot_layout()`, width and height of the output are adjusted to fit the content.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When TRUE (the default), ggsave will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{return_input}{If `TRUE` the input ggplot or plotlist is returned after saving.
This enables the use of `bro_ggsave_paged()` within `dplyr` pipes.}

\item{burst_to_multiple_files}{If `TRUE` multiple plots are saved to separate files instead of a single multi-page file.}

\item{...}{Other arguments passed on to the graphics device function, as specified by device.}
}
\description{
This function takes a ggplot (for single page) or list of ggplots (for multi page) and writes them to file.
In case the input has absolute dimensions (e.g. as a result of `egg::set_panel_size()` or
`patchwork::plot_layout()`), width and height of the output are adjusted to fit the content.
}
