% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_columns_to_rownames.R
\name{copy_column_to_rownames}
\alias{copy_column_to_rownames}
\title{Assign the rownames of a data frame to the values in a column without removing the column}
\usage{
copy_column_to_rownames(df_x, col_name)
}
\arguments{
\item{df_x}{A data frame}

\item{col_name}{The column name that contains rownames}
}
\value{
A data frame with row names
}
\description{
Assign the rownames of a data frame to the values in a column without removing the column
}
\examples{
library(mkhelpers)
car_df <- mtcars \%>\% rownames_to_column("car_name")
car_df <- car_df \%>\% copy_column_to_rownames("car_name")
}
