% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Title Filter on genes expressed in more than n samples (e.g. cells)}
\usage{
filter_genes(m, min_perc_exp, include_genes = "none")
}
\arguments{
\item{m}{A count matrix (can be a sparse matrix)}

\item{min_perc_exp}{The minimum percentage of samples expressing a given gene for it to be included in the output}

\item{include_genes}{A character vector of genes to include, even if their expression is lower than min_perc_exp}
}
\value{
A filtered count matrix
}
\description{
Title Filter on genes expressed in more than n samples (e.g. cells)
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
InstallData("pbmc3k")
seu_pbmc <- LoadData("pbmc3k", "pbmc3k.final")
filter_genes(m = pbmc3k@assays$RNA$counts, min_perc_exp = 5)
}
}
