% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chunks.R
\name{make_chunks}
\alias{make_chunks}
\title{Dividing a vector into chunks}
\usage{
make_chunks(x, n_chunks = 0, size_chunks = 1000)
}
\arguments{
\item{x}{A vector (e.g. a character vector or numeric vector)}

\item{n_chunks}{Set this to an integer to divide x into n chunks. Setting n_chunks to zero ignores the parameter}

\item{size_chunks}{Set this to the desired size of chunks and n_chunks will be automatically determined}
}
\value{
A list of chunked vectors
}
\description{
Dividing a vector into chunks
}
\examples{
char_vector <- purrr::map_chr(49:122, intToUtf8)
make_chunks(char_vector, n_chunks = 5)
make_chunks(char_vector, size_chunks = 3)
}
