% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkhelpers-package.R
\docType{package}
\name{mkhelpers-package}
\alias{mkhelpers}
\alias{mkhelpers-package}
\title{mkhelpers: My little helper functions}
\description{
A collection of my own functions that I frequently use in bioinformatic analyses.
}
\keyword{internal}
