% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_as_sorted_factor.R
\name{numbers_as_sorted_factor}
\alias{numbers_as_sorted_factor}
\title{Convert a character vector, factor or integer vector to an ordered factor preserving numerical order}
\usage{
numbers_as_sorted_factor(x)
}
\arguments{
\item{x}{An integer vector, a character vector (containing integer values as strings) or a factor (containing integer values as strings)}
}
\value{
An ordered factor with numerical order preserved
}
\description{
A common problem when converting numbers to factors is that the numeric order is not preserved. E.g. the numbers from 1 to 10 will be ordered as "1", "10", "2", "3", "4", "5", "6", "7", "8", "9". This function returns an ordered factor that preserves numerical order.
}
\examples{
\dontrun{
x <- factor(c("1", "2", "3", "10", "4", "10"))
levels(x)
numbers_as_sorted_factor(x)
numbers_as_sorted_factor(x) \%>\% levels

x <- c(1, 2, 3, 10, 4, 10)
numbers_as_sorted_factor(x)

x <- c("1", "2", "3", "10", "4", "10")
numbers_as_sorted_factor(x)
}
}
