% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_non_na.R
\name{perc_non_na}
\alias{perc_non_na}
\title{Report percentage of non-na values in tbl}
\usage{
perc_non_na(tbl_x)
}
\arguments{
\item{tbl_x}{A tibble}
}
\value{
A summarised tibble reporting the percentage of non-na values for all columns in the input tibble
}
\description{
Report percentage of non-na values in tbl
}
