% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_comparison.R
\name{plot_feature_comparison}
\alias{plot_feature_comparison}
\title{Title}
\usage{
plot_feature_comparison(
  tbl_x,
  title,
  expr_col,
  group_col,
  colors = "auto",
  output = c("image", "plot"),
  output_path,
  dpi = 600,
  point_size = 1.2,
  alpha = 0.8,
  plot_width = 25,
  plot_height = 15,
  out_width = 40,
  out_height = 30
)
}
\arguments{
\item{tbl_x}{A tibble containing feature expression values and group designations}

\item{title}{The plot title (e.g. the gene that is plotted)}

\item{expr_col}{The column, which contains the expression information}

\item{group_col}{The column, which designates the group (e.g. single cell clusters or treatment groups)}

\item{colors}{Here one can set the group colors manually by providing a named character vector (e.g. c(group_1 = "#666666", group_2 = "#FFC966"))}

\item{output}{Return a ggplot object ("plot") or save as a png image ("image").}

\item{output_path}{Output path.}

\item{dpi}{Resolution (dpi) for saved png.}

\item{point_size}{Point size passed to geom_jitter()}

\item{alpha}{Alpha passed to geom_jitter().}

\item{plot_width}{Width of the coordinate system in mm. Set to NA to leave undetermined.}

\item{plot_height}{Height of the coordinate system in mm. Set to NA to leave undetermined.}

\item{out_width}{Width in mm of the png output.}

\item{out_height}{Height in mm of the png output.}
}
\value{
A ggplot object (when output is set to "plot")
}
\description{
Title
}
\examples{
# See ?group_markers
}
