% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_comparison_grid.R
\name{plot_feature_comparison_grid}
\alias{plot_feature_comparison_grid}
\title{Compare a list of features across groups and samples with boxplots}
\usage{
plot_feature_comparison_grid(
  tbl_x,
  feature_list,
  feature_col = "gene",
  n_cols = 3,
  out_width = 180,
  output_path,
  ...
)
}
\arguments{
\item{tbl_x}{A tibble containing feature expression values and group designations}

\item{feature_list}{A list of features to be plotted from feature_col}

\item{feature_col}{The tibble column containing the feature names}

\item{n_cols}{Number of columns to plot}

\item{out_width}{Width of image output in mm}

\item{output_path}{Output path and file name (should end in .png)}

\item{...}{Other parameters passed on to plot_umap. Note, that expr_cola and group_col need to be provided.}
}
\description{
Compare a list of features across groups and samples with boxplots
}
