% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_mm.R
\name{plot_layout_mm}
\alias{plot_layout_mm}
\title{Wrapper for plot_layout with mm as default}
\usage{
plot_layout_mm(width, height, ...)
}
\arguments{
\item{width}{Plot width in mm}

\item{height}{Plot height in mm}

\item{...}{Parameters passed on to patchwork::plot_layout}
}
\value{
A patchwork
}
\description{
Wrapper for plot_layout with mm as default
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, cyl)) + geom_point()
p + plot_layout_mm(30, 30)
}
