% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Project discrete or continous features on a umap plot}
\usage{
plot_umap(
  tbl_x,
  umap_dim_col_1,
  umap_dim_col_2,
  feature_x,
  quantile_limits = c(0.3, 0.99),
  feature_colors = NULL,
  title,
  output = c("image", "plot"),
  output_path,
  dpi = 300,
  order_values = c("sorted", "random"),
  invert_sort_direction = FALSE,
  point_size = 0.3,
  point_size_legend = 1.5,
  alpha = 1,
  plot_width = 60,
  plot_height = 60,
  out_width = 89,
  out_height = 89,
  show_legend = TRUE,
  show_labels = FALSE,
  pretty_labels = FALSE,
  pretty_labels_precision = 50,
  label_size = 2
)
}
\arguments{
\item{tbl_x}{A tibble containing umap coordinates and feature columns}

\item{umap_dim_col_1}{Name of column containing umap dimension 1}

\item{umap_dim_col_2}{Name of column containing umap dimension 2}

\item{feature_x}{Name of column containing feature data}

\item{quantile_limits}{Quantile limits applied to continuous feature. E.g. c(0.1, 0.9) applies the dynamic range of the color scale only to to values above the 10th percentile and below the 90th percentile.}

\item{feature_colors}{For discrete features a color vector with one color per feature needs to be provided. A fixed assignment of colors to classes can be achieved by providing a named vector, where the vector names are the feature levels and the values the colors. For continous features provide either a vector of two or three colors to build a color gradient or leave this parameter empty to use the Cividis scale.}

\item{title}{Plot title.}

\item{output}{Return a ggplot object ("plot") or save as a png image ("image").}

\item{output_path}{Output path, shold include '.../imagename.png'.}

\item{dpi}{Resolution (dpi) for saved png.}

\item{order_values}{Whether to plot higher values in front of lower values ("sorted") or in random order ("random").}

\item{invert_sort_direction}{Whether to invert sort direction (plot lower values in front)}

\item{point_size}{Point size passed to geom_point().}

\item{alpha}{Alpha passed to geom_point().}

\item{plot_width}{Width of the coordinate system in mm. Set to NA to leave undetermined.}

\item{plot_height}{Height of the coordinate system in mm. Set to NA to leave undetermined.}

\item{out_width}{Width in mm of the png output.}

\item{out_height}{Height in mm of the png output.}

\item{show_legend}{Whether to show the ggplot legend.}

\item{show_labels}{Whether to label the mean coordinates of discrete feature values.}

\item{pretty_labels}{Whether to make labels pretty (takes longer)}

\item{pretty_labels_precision}{Tune the precision of label positioning. Higher values are more precise and take more processing time.}

\item{label_size}{Size of labels.}
}
\value{
A ggplot object (when output is set to "plot")
}
\description{
Project discrete or continous features on a umap plot
}
\examples{
\dontrun{
 library(mkhelpers)
 library(Seurat)
 library(SeuratData)
 InstallData("pbmc3k")
 seu_pbmc <- LoadData("pbmc3k", "pbmc3k.final")

 tbl_x <- seu_extract_tbl(seu_x = seu_pbmc,
   reduction = "umap",
   metadata_cols = "seurat_clusters",
   extract_expr = TRUE,
   genes_extract = "CD3D",
   assay_extract = "RNA",
   slot_extract = "data",
   expr_format = "wide")
 plot_umap(tbl_x = tbl_x, umap_dim_col_1 = UMAP_1, umap_dim_col_2 = UMAP_2, output_path = "~", title = "Seurat clusters", feature_x = seurat_clusters, order_values = "random", show_legend = FALSE, show_labels = TRUE, point_size = 0.5)
 plot_umap(tbl_x = tbl_x, umap_dim_col_1 = UMAP_1, umap_dim_col_2 = UMAP_2, output_path = "~", title = "CD3 expression", feature_x = CD3D, point_size = 0.5)
}
}
