% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_na.R
\name{rm_na}
\alias{rm_na}
\title{Remove NAs from vector or list}
\usage{
rm_na(x)
}
\arguments{
\item{A}{vector or list}
}
\value{
A vector or list with NA entries removed
}
\description{
Remove NAs from vector or list
}
\examples{
x <- c("something", "something else", NA, "and another thing", NA)
rm_na(x)
}
