% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_list.R
\name{row_list}
\alias{row_list}
\title{Split a dataframe into a list, where each entry is one row of the dataframe}
\usage{
row_list(df)
}
\arguments{
\item{df}{The input dataframe}
}
\value{
A list of rows
}
\description{
Split a dataframe into a list, where each entry is one row of the dataframe
}
