% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seu_extract_tbl.R
\name{seu_extract_tbl}
\alias{seu_extract_tbl}
\title{Extracting information stored in seurat objects into a tibble}
\usage{
seu_extract_tbl(
  seu_x,
  reduction = NULL,
  metadata_cols = "all",
  extract_expr = FALSE,
  genes_extract = "all",
  assay_extract = "RNA",
  slot_extract = "data",
  expr_format = c("wide", "long")
)
}
\arguments{
\item{seu_x}{Seurat object}

\item{reduction}{Name of reduction to be extracted}

\item{metadata_cols}{Names of metadata columns to be extracted}

\item{extract_expr}{Whether to extract expression information (default: FALSE)}

\item{genes_extract}{Genes to extract, when extracting expression information (default: "all")}

\item{assay_extract}{Assay to extract expression information from (default: "RNA")}

\item{slot_extract}{Slot to extract expression information from (default: "data")}

\item{expr_format}{Return gene expression information in wide or long format (default: "wide")}
}
\value{
A tibble
}
\description{
Extracting information stored in seurat objects into a tibble
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
InstallData("pbmc3k")
seu_pbmc <- LoadData("pbmc3k", "pbmc3k.final")

seu_extract_tbl(seu_x = seu_pbmc,
  reduction = "umap",
  metadata_cols = c("nCount_RNA", "nFeature_RNA", "seurat_clusters"),
  extract_expr = TRUE,
  genes_extract = c("NCAM1", "SELL", "CD3D"),
  assay_extract = "RNA",
  slot_extract = "data",
  expr_format = "wide")
}
}
