% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_select_fields.R
\name{str_select_fields}
\alias{str_select_fields}
\title{Select fields in a string}
\usage{
str_select_fields(x, sep, fields, leading_sep = FALSE, trailing_sep = FALSE)
}
\arguments{
\item{x}{A string to select fields from}

\item{sep}{The separator between fields}

\item{fields}{Which fields to choose, e.g. c(1,2,5) or 1:6}

\item{leading_sep}{Whether to add/preserve a leading separator}

\item{trailing_sep}{Whether to add/preserve a trailing separator}
}
\value{
A string
}
\description{
Select fields in a string
}
\examples{
library(mkhelpers)
x <- "apple-banana-pear-mango-peach"
str_select_fields(x, "-", c(1,5))
}
