% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_add_percentiles.R
\name{tbl_add_percentiles}
\alias{tbl_add_percentiles}
\title{Add percentile column to tibble}
\usage{
tbl_add_percentiles(tbl_x, var_x)
}
\arguments{
\item{tbl_x}{A tibble}

\item{var_x}{The tibble column for which to calculate percentiles}
}
\value{
A tibble
}
\description{
Adds a new column to the tibble, which shows the percentiles for a column of choice
}
\examples{
library(mkhelpers)
tbl_add_percentiles(mtcars, disp)
}
