% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mk.R
\docType{data}
\name{theme_mk}
\alias{theme_mk}
\title{Costum ggplot2 theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 94.
}
\usage{
theme_mk
}
\description{
Costum ggplot2 theme
}
\keyword{datasets}
