% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_priority.R
\name{tidyverse_priority}
\alias{tidyverse_priority}
\title{Giving all tidyverse functions priority over conflicting functions from other packages}
\usage{
tidyverse_priority()
}
\description{
Giving all tidyverse functions priority over conflicting functions from other packages
}
\examples{
library(mkhelpers)
library(tidyverse)
library(GenomicRanges)
tidyverse_priority()
}
