% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_preview.R
\name{color_preview}
\alias{color_preview}
\title{Previews a vector of hexadecimal colors}
\usage{
color_preview(color_vector)
}
\arguments{
\item{color_vector}{A hexadecimal color vector to plot}
}
\description{
Previews a vector of hexadecimal colors
}
\examples{
library(mkhelpers)
color_preview(c("#E04B4B", "#EC9393", "#3B537F", "#8998B2", "#969696", "#F2BF44"))
}
